# packages/shared/openapi.yaml
openapi: 3.1.0
info:
  title: rescue-net.eu API
  version: 0.1.0
  description: >-
    Volunteer rescue alerting and mission coordination platform.
    This is the MVP skeleton contract. Endpoints are defined in the project
    manual section 14; only the implemented surface is documented here and the
    file grows as features land.
servers:
  - url: https://api.rescue-net.eu
  - url: http://localhost:8000
paths:
...
  /missions:
    get:
      summary: List missions (membership-scoped for non-staff)
      tags: [missions]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Mission collection
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Mission" }
  /missions/{mission_id}:
    get:
      summary: Get a mission
      tags: [missions]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: mission_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Mission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mission" }
        "403": { description: Not a mission member }
    patch:
      summary: Update mission status (staff or Team Lead; manual §8)
      tags: [missions]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: mission_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, mobilizing, active, waiting, returning, closed, cancelled]
      responses:
        "200":
          description: Updated mission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mission" }
  /missions/{mission_id}/join:
    post:
      summary: Join a mission; enable live location with explicit consent (§5.3, §16.2)
      tags: [missions]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: mission_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                live_location_enabled: { type: boolean, default: false }
      responses:
        "200":
          description: Updated mission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mission" }
  /missions/{mission_id}/leave:
    post:
      summary: Leave a mission; stops live location sharing (§16.2)
      tags: [missions]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: mission_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Updated mission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mission" }
  /missions/{mission_id}/close:
    post:
      summary: Close a mission (staff or Team Lead; §5.5)
      tags: [missions]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: mission_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                closure_summary: { type: string, nullable: true }
      responses:
        "200":
          description: Closed mission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Mission" }
  /missions/{mission_id}/messages:
    get:
      summary: List mission chat messages
      tags: [missions]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: mission_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChatMessage" }
    post:
      summary: Post a mission chat message
      tags: [missions]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: mission_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string }
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatMessage" }
        "409": { description: Mission is closed }
  /missions/{mission_id}/locations:
    post:
      summary: Ingest a live location sample (active member with consent; §16.2)
      tags: [missions]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: mission_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude]
              properties:
                latitude: { type: number }
                longitude: { type: number }
                accuracy_m: { type: number, nullable: true }
                speed: { type: number, nullable: true }
                heading: { type: number, nullable: true }
      responses:
        "201":
          description: Stored location
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Location" }
        "403": { description: Not an active member or consent not given }
...
  /missions/{mission_id}/tasks:
    get:
      summary: List mission tasks (members; §5.4)
      tags: [missions]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: mission_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
    post:
      summary: Create/assign a task (staff or Team Lead; §6.4)
      tags: [missions]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: mission_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string, nullable: true }
                priority: { type: string, enum: [low, medium, high], default: medium }
                assigned_to: { type: string, format: uuid, nullable: true }
                due_at: { type: string, format: date-time, nullable: true }
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "403": { description: Only staff or the Team Lead may create tasks }
        "422": { description: Assignee is not an active mission member }
  /missions/{mission_id}/tasks/{task_id}:
    patch:
      summary: Update a task (staff/lead any field; assignee may set own status; §5.4)
      tags: [missions]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: mission_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: task_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string, nullable: true }
                status: { type: string, enum: [open, in_progress, blocked, done, cancelled] }
                priority: { type: string, enum: [low, medium, high] }
                assigned_to: { type: string, format: uuid, nullable: true }
                due_at: { type: string, format: date-time, nullable: true }
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "403": { description: Responders may only set the status of their own task }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MissionMember:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        role_in_mission: { type: string }
        joined_at: { type: string, format: date-time }
        left_at: { type: string, format: date-time, nullable: true }
        live_location_enabled: { type: boolean }
    Mission:
      type: object
      properties:
        id: { type: string, format: uuid }
        incident_id: { type: string, format: uuid }
        lead_user_id: { type: string, format: uuid, nullable: true }
        status:
          type: string
          enum: [pending, mobilizing, active, waiting, returning, closed, cancelled]
        started_at: { type: string, format: date-time, nullable: true }
        closed_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        members:
          type: array
          items: { $ref: "#/components/schemas/MissionMember" }
    ChatMessage:
      type: object
      properties:
        id: { type: string, format: uuid }
        mission_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        message: { type: string }
        created_at: { type: string, format: date-time }
    Location:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        latitude: { type: number }
        longitude: { type: number }
        accuracy_m: { type: number, nullable: true }
        speed: { type: number, nullable: true }
        heading: { type: number, nullable: true }
        timestamp: { type: string, format: date-time }
    Task:
      type: object
      properties:
        id: { type: string, format: uuid }
        mission_id: { type: string, format: uuid }
        assigned_to: { type: string, format: uuid, nullable: true }
        created_by: { type: string, format: uuid }
        title: { type: string }
        description: { type: string, nullable: true }
        status: { type: string, enum: [open, in_progress, blocked, done, cancelled] }
        priority: { type: string, enum: [low, medium, high] }
        due_at: { type: string, format: date-time, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
...